refactor(Halls): refactored logger, error handlers and function arguments

+++ internal/interfaces.go
@@ -24,5 +24,5 @@ type Service interface {
 }
 
 type Identifiable interface {
-	GID() int
+	GID() int64
 }
