+++ internal/service/halls/service.go
@@ -26,7 +26,7 @@ func (s *Service) Create(r internal.Identifiable) (internal.Identifiable, error)
 
 // Retrieve logic layer for repository method
 func (s *Service) Retrieve(id int64) (internal.Identifiable, error) {
-	return s.repo.Retrieve(int64(id))
+	return s.repo.Retrieve(id)
 }
 
 // RetriveAll logic layer for repository method
@@ -36,5 +36,5 @@ func (s *Service) RetrieveAll() ([]internal.Identifiable, error) {
 
 // Delete logic layer for repository method
 func (s *Service) Delete(id int64) error {
-	return s.repo.Delete(int64(id))
+	return s.repo.Delete(id)
 }
